examples/dtls-echo: Cleaning and upgrade to sock
Due to all the changes, this is basically a new version for this
example. The main benefit is the use of sock_udp but also the client
side is now more robust and reliable.

The parameters required for the PSK and ECC (e.g. keys) modes are moved
to an unique header.